+++ src/js/index.ts
@@ -57,7 +57,12 @@ export { ReactNativeClient } from "./client";
 export { init, setDist, setRelease, nativeCrash } from "./sdk";
 export { TouchEventBoundary, withTouchEventBoundary } from "./touchevents";
 
-export * as Tracing from "./tracing";
+export {
+  ReactNativeTracing,
+  ReactNavigationV4Instrumentation,
+  ReactNavigationV5Instrumentation,
+  RoutingInstrumentation,
+} from "./tracing";
 
 /**
  * Adds the sdk info. Make sure this is called after @sentry/react's so this is the top-level SDK.
