+++ src/js/tracing/index.ts
@@ -1,4 +1,7 @@
 export { ReactNativeTracing } from "./reactnativetracing";
 export { ReactNavigationV5Instrumentation } from "./reactnavigationv5";
 export { ReactNavigationV4Instrumentation } from "./reactnavigationv4";
-export { RoutingInstrumentation } from "./router";
+export {
+  RoutingInstrumentation,
+  RoutingInstrumentationInstance,
+} from "./routingInstrumentation";
