+++ src/js/tracing/reactnavigationv5.ts
@@ -1,7 +1,7 @@
 import { Transaction as TransactionType } from "@sentry/types";
 import { logger } from "@sentry/utils";
 
-import { RoutingInstrumentation } from "./router";
+import { RoutingInstrumentation } from "./routingInstrumentation";
 interface NavigationRoute {
   name: string;
   key: string;
